feat: implement tool registration orchestration for MCP implementations
